Renamed annotation + annotationdatasetmodel
preparing merge of annotation and MVector
