+++ app/addselectiondialog.h
@@ -4,7 +4,7 @@
 #include <QtCore>
 #include <QDialog>
 #include <QMap>
-#include "annotation.h"
+#include "annotation_old.h"
 #include "mvector.h"
 #include <QTableWidgetItem>
