+++ app/editannotationdatawindow.h
@@ -3,7 +3,7 @@
 
 #include <QtCore>
 #include <QMainWindow>
-#include "annotationdatasetmodel.h"
+#include "annotationdatasetmodel_old.h"
 #include "mvector.h"
 
 namespace Ui {
